pyproject.toml
[tool.poetry]
name = "quote-manager-cli"
version = "0.1.0"
description = "A CLI tool for managing quotes, including adding, listing, generating random quotes, and initializing the database with quotes from a JSON file."

authors = ["Prechy <preciousrobinsonokafor@gmail.com>"]

readme = "README.md"

repository = "https://github.com/Data-Epic/Quote-Manager-CLI-Precious/"

# Keywords 
keywords = ["Quote Manager", "CLI Tool"]

[tool.poetry.dependencies]
python = "^3.8.1"
duckdb-engine = "^0.13.1"
sqlalchemy = "^2.0.32"
click = "^8.1.7"
python-dotenv = "^1.0.1"

[tool.poetry.group.test.dependencies]
pytest = "^8.3.2"


[tool.poetry.group.dev.dependencies]
isort = "^5.13.2"
mypy = "^1.11.2"
ruff = "^0.6.4"

[tool.ruff]
line-length = 100

[tool.poetry.scripts]
quote = "quote_manager_cli.cli:cli"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
